charger: allow devices to skip the initial display blank
The theory behind the blank was that the kernel would not
re-initilize the display on boot (and leave it to whatever
the bootloader configured it to) until after the first
blank request.

On some devices, the bootloader does not turn the screen on,
but it's done by the kernel and thus. This makes the extra blank at
start unnecessary.

Bug: 7319529
Change-Id: I9ad9856186df6e9e23ebbe86e83b641423e16dcb
Signed-off-by: Dima Zavin <dima@android.com>