warn_collector: handle non-fatal signals
If we pause the process (standard SIGSTOP/SIGCONT), the read() on the
inotify handle gets angry and aborts.  This can happen if you ctrl+z
the program, or if you attach to it with strace.  This is undesirable,
so ignore EINTR errors from the read().

BUG=None
TEST=ran `./warn_collector --test` and hit ctrl+z then fg

Change-Id: I49e99abb12ace847e88f973817854d8ba95cf689
Reviewed-on: https://gerrit.chromium.org/gerrit/56303
Reviewed-by: Luigi Semenzato <semenzato@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Queue: Mike Frysinger <vapier@chromium.org>