Don't set LD_LIBRARY_PATH globally.
On a 64-bit system, 64-bit processes will want one path, 32-bit processes
another. The dynamic linker already provides the correct defaults for
native code, and we've coupled the VM and dynamic linker so that
LD_LIBRARY_PATH will be set correctly in any VM during startup if it's not
being manually overridden.

Change-Id: Icbffc0d451dbc242cdfb9267413d8bcac434e108