EA-10: Allow symlink masks with minor number or bInterfaceNumber
This is an improvement on the previous implementation as it does not
force the major/minor to be specified if it's not needed. Only the
Vendor/Product IDs are really required. If not present, then a symlink
will be created for each of the endpoints that has to be created for
the USB device. To avoid name conflicts, the alias name can include %i
or %m has mask. %i will be replaced by the bInterfaceNumber of the
endpoint, and %m will be replaced by the minor number.

Change-Id: I3f75c68664a39860279c8d5b64d6cfa6062c8e00