EA-301: Properly remove links using bInterfaceNumber in masks
It is impossible to obtain the bInterfaceNumber of devices being
removed. We only have major/minor, product ID and vendor ID for the
device that is going. The resolution to this problem was not obvious as
ueventd is entirely stateless. We device to save the state that would
be kept inside ueventd by creating symlinks to the alias in the
/dev/aliases directory. We use the major/minor number of the device to
which the alias point for the name of the symlink pointing to the
alias. That way, we can easily remove aliases that were created for the
device when we receive the major/minor numbers of the device being
removed.