Increase sysprop timeout
Change the sysprop socket timeout from 2000ms to 5000ms.  This has no
effect when there are no timeouts.  However, this improves reliability
in lower-end systems are under heavy system load.

This has no effect on sysprop errors that are not timeouts.  In
particular, it does not affect the timing when there are permission
errors or attempts to update a read-only property (both of these
errors are seen quite frequently in logs).

Flag: EXEMPT bug-fix
Bug: 383751329
Test: presubmit
Change-Id: Ibc70db1f6c565e0a8651d9026854d66db36418dd